# planner: avoid repeat normalization on the same query when matching binding for plan cache