--- a/pkg/bindinfo/binding_match.go
+++ b/pkg/bindinfo/binding_match.go
@@ -15,6 +15,7 @@
 package bindinfo
 
 import (
+	"github.com/pingcap/tidb/pkg/util/hint"
 	"strings"
 	"sync"
 
@@ -31,38 +32,72 @@ var (
 	GetGlobalBindingHandle func(sctx sessionctx.Context) GlobalBindingHandle
 )
 
+// BindingMatchInfo records necessary information for fuzzy binding matching.
+// This is mainly for plan cache to avoid normalizing the same statement repeatedly.
+type BindingMatchInfo struct {
+	FuzzyDigest string
+	TableNames  []*ast.TableName
+}
+
+// MatchSQLBindingForPlanCache matches binding for plan cache.
+func MatchSQLBindingForPlanCache(sctx sessionctx.Context, stmtNode ast.StmtNode, info *BindingMatchInfo) (bindingSQL string, ignoreBinding bool) {
+	bindRecord, _ := getBindRecord(sctx, stmtNode, info)
+	if bindRecord != nil {
+		if enabledBinding := bindRecord.FindEnabledBinding(); enabledBinding != nil {
+			bindingSQL = enabledBinding.BindSQL
+			ignoreBinding = enabledBinding.Hint.ContainTableHint(hint.HintIgnorePlanCache)
+		}
+	}
+	return
+}
+
 // MatchSQLBinding returns the matched binding for this statement.
 func MatchSQLBinding(sctx sessionctx.Context, stmtNode ast.StmtNode) (bindRecord *BindRecord, scope string, matched bool) {
-	useBinding := sctx.GetSessionVars().UsePlanBaselines
-	if !useBinding || stmtNode == nil {
-		return nil, "", false
-	}
-	var err error
-	bindRecord, scope, err = getBindRecord(sctx, stmtNode)
-	if err != nil || bindRecord == nil || len(bindRecord.Bindings) == 0 {
+	bindRecord, scope = getBindRecord(sctx, stmtNode, nil)
+	if bindRecord == nil || len(bindRecord.Bindings) == 0 {
 		return nil, "", false
 	}
 	return bindRecord, scope, true
 }
 
-func getBindRecord(ctx sessionctx.Context, stmt ast.StmtNode) (*BindRecord, string, error) {
+func getBindRecord(sctx sessionctx.Context, stmtNode ast.StmtNode, info *BindingMatchInfo) (*BindRecord, string) {
+	useBinding := sctx.GetSessionVars().UsePlanBaselines
+	if !useBinding || stmtNode == nil {
+		return nil, ""
+	}
 	// When the domain is initializing, the bind will be nil.
-	if ctx.Value(SessionBindInfoKeyType) == nil {
-		return nil, "", nil
+	if sctx.Value(SessionBindInfoKeyType) == nil {
+		return nil, ""
 	}
+
+	// record the normalization result into info to avoid repeat normalization next time.
+	var fuzzyDigest string
+	var tableNames []*ast.TableName
+	if info == nil || info.TableNames == nil || info.FuzzyDigest == "" {
+		_, fuzzyDigest = NormalizeStmtForFuzzyBinding(stmtNode)
+		tableNames = CollectTableNames(stmtNode)
+		if info != nil {
+			info.FuzzyDigest = fuzzyDigest
+			info.TableNames = tableNames
+		}
+	} else {
+		fuzzyDigest = info.FuzzyDigest
+		tableNames = info.TableNames
+	}
+
 	// the priority: session normal > session universal > global normal > global universal
-	sessionHandle := ctx.Value(SessionBindInfoKeyType).(SessionBindingHandle)
-	if bindRecord, err := sessionHandle.MatchSessionBinding(ctx, stmt); err == nil && bindRecord != nil && bindRecord.HasEnabledBinding() {
-		return bindRecord, metrics.ScopeSession, nil
+	sessionHandle := sctx.Value(SessionBindInfoKeyType).(SessionBindingHandle)
+	if bindRecord, err := sessionHandle.MatchSessionBinding(sctx, fuzzyDigest, tableNames); err == nil && bindRecord != nil && bindRecord.HasEnabledBinding() {
+		return bindRecord, metrics.ScopeSession
 	}
-	globalHandle := GetGlobalBindingHandle(ctx)
+	globalHandle := GetGlobalBindingHandle(sctx)
 	if globalHandle == nil {
-		return nil, "", nil
+		return nil, ""
 	}
-	if bindRecord, err := globalHandle.MatchGlobalBinding(ctx, stmt); err == nil && bindRecord != nil && bindRecord.HasEnabledBinding() {
-		return bindRecord, metrics.ScopeGlobal, nil
+	if bindRecord, err := globalHandle.MatchGlobalBinding(sctx, fuzzyDigest, tableNames); err == nil && bindRecord != nil && bindRecord.HasEnabledBinding() {
+		return bindRecord, metrics.ScopeGlobal
 	}
-	return nil, "", nil
+	return nil, ""
 }
 
 func eraseLastSemicolon(stmt ast.StmtNode) {