--- a/pkg/executor/prepared.go
+++ b/pkg/executor/prepared.go
@@ -16,10 +16,10 @@ package executor
 
 import (
 	"context"
-	"github.com/pingcap/tidb/pkg/bindinfo"
 
 	"github.com/pingcap/errors"
 	"github.com/pingcap/log"
+	"github.com/pingcap/tidb/pkg/bindinfo"
 	"github.com/pingcap/tidb/pkg/executor/internal/exec"
 	"github.com/pingcap/tidb/pkg/expression"
 	"github.com/pingcap/tidb/pkg/infoschema"