… pkg/bindinfo/capture_test.go