<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: ref #48875

Problem Summary: planner: avoid repeat normalization on the same query when matching binding for plan cache

### What changed and how does it work?

planner: avoid repeat normalization on the same query when matching binding for plan cache

Before:
```
BenchmarkPlanCacheBindingMatch:
   21859             55975 ns/op           22666 B/op        360 allocs/op
   21502             56296 ns/op           22704 B/op        361 allocs/op
   20362             56433 ns/op           22714 B/op        361 allocs/op
```

After
```
BenchmarkPlanCacheBindingMatch:
   23390             54306 ns/op           22055 B/op        345 allocs/op
   21025             54262 ns/op           22046 B/op        345 allocs/op
   19862             54507 ns/op           21991 B/op        345 allocs/op
```

### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [ ] Integration test
- [ ] Manual test (add detailed scripts or steps below)
- [ ] No need to test
  > - [ ] I checked and no code files have been changed.
  > <!-- Or your custom  "No need to test" reasons -->

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- compatibility change, improvement, bugfix, and new feature need a release note -->

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

```release-note
None
```